Refactor command classes
In our command classes, execute() returns a string for the Ui to print.

This is unintuitive and conveys no information about the success or
failure of the execution action.

Let's have execute() return its result status instead.

This would make it easier to write tests and assertions, aligning with
our goals for a more robust application.