feat: Add loading data from JSON functionality
The loading of the data happens automatically when the program
first starts by calling `FileReaderWriter.loadMediTrackerData`.

Some error handling is provided to ensure that the program does not
crash when starting up for the first time, or when reading corrupt JSON
save file, and instead treat them as invalid and the program will run
with a clean slate.